Create License.txt
Copyright (c) Apache 2.0 2023

Permission is hereby granted, free of charge, to any end-users for obtaining a copy of source code, for their own proprieary software, to alter source code, update source code, and to distribute modifications and copies of the source code provided in this repository. All altered files must contain an Apache 2.0 license with attribution. Additional copyright statements to modified and distributed copies of source code is permitted. End-users who modify and distribute modifications of source code can create their own terms of use.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.